fix: change tab's behavior
Before when you would close the last tab, the active tab index wouldn't change. Now, however, when you close the last tab the active tab index decrements by one. Added , a boolean value, which enables or disables vimi to focus on editors. This value is used to prevent vimi from focusing on the active tab's editor when a new tab is opened.